@@ -1,21 +1,6 @@
 apply plugin: 'com.android.application'
 
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-
-    dependencies {
-        classpath 'me.tatarka:gradle-retrolambda:2.4.0'
-    }
-}
-
-// Required because retrolambda is on maven central
-repositories {
-    mavenCentral()
-}
-
-apply plugin: 'retrolambda'
+apply plugin: 'me.tatarka.retrolambda'
 
 android {
     signingConfigs {
@@ -26,14 +11,14 @@ android {
             storePassword System.getenv("SIGN_PASSWORD")
         }
     }
-    compileSdkVersion 21
-    buildToolsVersion '21.1.2'
+    compileSdkVersion 22
+    buildToolsVersion '22.0.1'
     defaultConfig {
         applicationId 'yuku.alkitab.debug'
         minSdkVersion 14
-        targetSdkVersion 21
-        versionCode 14000179
-        versionName '4.0.0-beta14'
+        targetSdkVersion 22
+        versionCode 14000202
+        versionName '4.1.0-beta2'
         multiDexEnabled true
     }
     buildTypes {
@@ -77,14 +62,17 @@ dependencies {
     compile project(':AmbilWarna')
     compile project(':FileChooser')
     compile project(':ImportedDesktopVerseUtil')
-    compile 'com.android.support:support-v4:21.0.3'
+    compile 'com.android.support:support-v4:22.1.1'
     compile 'com.squareup.picasso:picasso:2.4.0'
     compile 'com.squareup.okhttp:okhttp:2.1.0'
     compile 'com.squareup.okhttp:okhttp-urlconnection:2.1.0'
     compile 'net.sf.trove4j:trove4j:3.0.3'
     compile 'com.google.code.gson:gson:2.3.1'
     compile 'com.android.support:support-v13:20.0.0'
-    compile 'com.google.android.gms:play-services-base:6.5.87'
-    compile 'com.android.support:appcompat-v7:21.0.3'
+    compile 'com.google.android.gms:play-services-analytics:7.3.0'
+    compile 'com.google.android.gms:play-services-gcm:7.3.0'
+    compile 'com.android.support:appcompat-v7:22.1.1'
     compile 'com.android.support:multidex:1.0.0'
+    compile 'com.afollestad:material-dialogs:0.7.3.0'
+    testCompile 'junit:junit:4.12'
 }
