@@ -16,7 +16,6 @@
 
 package com.google.samples.apps.iosched.ui.widget;
 
-import android.R;
 import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.Color;
@@ -58,7 +57,7 @@ class SlidingTabStrip extends LinearLayout {
 		final float density = getResources().getDisplayMetrics().density;
 
 		TypedValue outValue = new TypedValue();
-		context.getTheme().resolveAttribute(R.attr.colorForeground, outValue, true);
+		context.getTheme().resolveAttribute(android.R.attr.colorForeground, outValue, true);
 		final int themeForegroundColor =  outValue.data;
 
 		mDefaultBottomBorderColor = setColorAlpha(themeForegroundColor,
