@@ -1,6 +1,5 @@
 package yuku.alkitab.base.ac;
 
-import android.app.Activity;
 import android.appwidget.AppWidgetManager;
 import android.content.BroadcastReceiver;
 import android.content.ComponentName;
@@ -10,27 +9,25 @@ import android.content.IntentFilter;
 import android.os.Bundle;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.CheckBox;
 import android.widget.CheckedTextView;
-import android.widget.CompoundButton;
 import android.widget.ListView;
 import android.widget.SeekBar;
 import android.widget.TextView;
 import yuku.afw.V;
-import yuku.afw.storage.Preferences;
 import yuku.afw.widget.EasyAdapter;
 import yuku.alkitab.base.App;
 import yuku.alkitab.base.S;
+import yuku.alkitab.base.ac.base.BaseActivity;
 import yuku.alkitab.base.br.DailyVerseAppWidgetReceiver;
 import yuku.alkitab.base.model.MVersion;
+import yuku.alkitab.base.util.DailyVerseData;
 import yuku.alkitab.debug.R;
 
 import java.util.List;
 
-public class DailyVerseAppWidgetConfigurationActivity extends Activity {
-	private static final String TAG = DailyVerseAppWidgetConfigurationActivity.class.getSimpleName();
+public class DailyVerseAppWidgetConfigurationActivity extends BaseActivity {
 	private VersionAdapter adapter;
 
 	int mAppWidgetId = AppWidgetManager.INVALID_APPWIDGET_ID;
@@ -38,6 +35,9 @@ public class DailyVerseAppWidgetConfigurationActivity extends Activity {
 	CheckBox cDarkText;
 	SeekBar sbTextSize;
 	TextView tTextSize;
+	View panelTransparent;
+	SeekBar sbTransparent;
+	TextView tTransparent;
 	private CheckBox cTransparentBackground;
 
 	final BroadcastReceiver br = new BroadcastReceiver() {
@@ -63,13 +63,15 @@ public class DailyVerseAppWidgetConfigurationActivity extends Activity {
 		cDarkText = V.get(this, R.id.cDarkText);
 		sbTextSize = V.get(this, R.id.sbTextSize);
 		tTextSize = V.get(this, R.id.tTextSize);
+		panelTransparent = V.get(this, R.id.panelTransparent);
+		sbTransparent = V.get(this, R.id.sbTransparent);
+		tTransparent = V.get(this, R.id.tTransparent);
 
 		// Find the widget id from the intent.
-		Intent intent = getIntent();
-		Bundle extras = intent.getExtras();
+		final Intent intent = getIntent();
+		final Bundle extras = intent.getExtras();
 		if (extras != null) {
-			mAppWidgetId = extras.getInt(
-			AppWidgetManager.EXTRA_APPWIDGET_ID, AppWidgetManager.INVALID_APPWIDGET_ID);
+			mAppWidgetId = extras.getInt(AppWidgetManager.EXTRA_APPWIDGET_ID, AppWidgetManager.INVALID_APPWIDGET_ID);
 		}
 
 		// If they gave us an intent without the widget id, just bail.
@@ -78,26 +80,26 @@ public class DailyVerseAppWidgetConfigurationActivity extends Activity {
 			return;
 		}
 
-		String key = "app_widget_" + mAppWidgetId + "_click";
-		Preferences.setInt(key, 0);
-
 		adapter = new VersionAdapter();
 		adapter.reload();
 		lsVersionsAppWidget.setAdapter(adapter);
-		lsVersionsAppWidget.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-			@Override
-			public void onItemClick(final AdapterView<?> parent, final View view, final int position, final long id) {
-				selectedVersionPosition = position;
-				bOk.setEnabled(true);
-				adapter.notifyDataSetChanged();
-			}
+		lsVersionsAppWidget.setOnItemClickListener((parent, view, position, id) -> {
+			selectedVersionPosition = position;
+			bOk.setEnabled(true);
+			adapter.notifyDataSetChanged();
 		});
 
 		bOk.setEnabled(false);
 		bOk.setOnClickListener(bOk_click);
-		bCancel.setOnClickListener(bCancel_click);
+		bCancel.setOnClickListener(v -> finish());
 
-		cTransparentBackground.setOnCheckedChangeListener(cTransparentBackground_checkedChange);
+		cTransparentBackground.setOnCheckedChangeListener((buttonView, isChecked) -> {
+			panelTransparent.setVisibility(isChecked ? View.VISIBLE : View.GONE);
+			cDarkText.setEnabled(isChecked);
+			if (!isChecked) {
+				cDarkText.setChecked(false);
+			}
+		});
 
 		sbTextSize.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
 			@Override
@@ -106,12 +108,30 @@ public class DailyVerseAppWidgetConfigurationActivity extends Activity {
 			}
 
 			@Override
-			public void onStartTrackingTouch(final SeekBar seekBar) {}
+			public void onStartTrackingTouch(final SeekBar seekBar) {
+			}
 
 			@Override
-			public void onStopTrackingTouch(final SeekBar seekBar) {}
+			public void onStopTrackingTouch(final SeekBar seekBar) {
+			}
 		});
 		sbTextSize_progressChanged(sbTextSize.getProgress());
+
+		sbTransparent.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
+			@Override
+			public void onProgressChanged(final SeekBar seekBar, final int progress, final boolean fromUser) {
+				sbTransparent_progressChanged(progress);
+			}
+
+			@Override
+			public void onStartTrackingTouch(final SeekBar seekBar) {
+			}
+
+			@Override
+			public void onStopTrackingTouch(final SeekBar seekBar) {
+			}
+		});
+		sbTransparent_progressChanged(sbTransparent.getProgress());
 
 		App.getLbm().registerReceiver(br, new IntentFilter(VersionsActivity.VersionListFragment.ACTION_RELOAD));
 	}
@@ -128,6 +148,19 @@ public class DailyVerseAppWidgetConfigurationActivity extends Activity {
 		tTextSize.setText("" + (int) textSize);
 	}
 
+	void sbTransparent_progressChanged(final int progress) {
+		final int percent = progressToActualTransparentPercent(progress);
+		tTransparent.setText(percent + "%");
+	}
+
+	int progressToActualTransparentPercent(final int progress) {
+		return progress * 5;
+	}
+
+	int progressToActualAlpha(final int progress) {
+		return (int) (255.f * ((100 - progressToActualTransparentPercent(progress)) / 100.f));
+	}
+
 	float progressToActualTextSize(final int progress) {
 		return progress + 8.f;
 	}
@@ -136,20 +169,18 @@ public class DailyVerseAppWidgetConfigurationActivity extends Activity {
 		@Override
 		public void onClick(final View v) {
 			final Context context = DailyVerseAppWidgetConfigurationActivity.this;
-			AppWidgetManager appWidgetManager = AppWidgetManager.getInstance(context);
-			String versionId = adapter.versions.get(selectedVersionPosition).getVersionId();
+			final String versionId = adapter.versions.get(selectedVersionPosition).getVersionId();
 
-			Preferences.hold();
-			try {
-				Preferences.setString("app_widget_" + mAppWidgetId + "_version", versionId);
-				Preferences.setBoolean("app_widget_" + mAppWidgetId + "_option_transparent_background", cTransparentBackground.isChecked());
-				Preferences.setBoolean("app_widget_" + mAppWidgetId + "_option_dark_text", cDarkText.isChecked());
-				Preferences.setFloat("app_widget_" + mAppWidgetId + "_option_text_size", progressToActualTextSize(sbTextSize.getProgress()));
-			} finally {
-				Preferences.unhold();
-			}
+			final DailyVerseData.SavedState savedState = new DailyVerseData.SavedState();
+			savedState.versionId = versionId;
+			savedState.transparentBackground = cTransparentBackground.isChecked();
+			savedState.backgroundAlpha = cTransparentBackground.isChecked() ? progressToActualAlpha(sbTransparent.getProgress()) : 255;
+			savedState.darkText = cDarkText.isChecked();
+			savedState.textSize = progressToActualTextSize(sbTextSize.getProgress());
+			savedState.click = 0;
+			DailyVerseData.saveSavedState(mAppWidgetId, savedState);
 
-			DailyVerseAppWidgetReceiver.buildUpdate(context, appWidgetManager, mAppWidgetId);
+			DailyVerseAppWidgetReceiver.buildUpdate(context, mAppWidgetId, 1);
 
 			ComponentName provider = new ComponentName(context, DailyVerseAppWidgetReceiver.class);
 			int[] ids = AppWidgetManager.getInstance(context).getAppWidgetIds(provider);
@@ -159,23 +190,6 @@ public class DailyVerseAppWidgetConfigurationActivity extends Activity {
 			resultValue.putExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, mAppWidgetId);
 			setResult(RESULT_OK, resultValue);
 			finish();
-		}
-	};
-
-	private View.OnClickListener bCancel_click = new View.OnClickListener() {
-		@Override
-		public void onClick(final View v) {
-			finish();
-		}
-	};
-
-	private CompoundButton.OnCheckedChangeListener cTransparentBackground_checkedChange = new CompoundButton.OnCheckedChangeListener() {
-		@Override
-		public void onCheckedChanged(final CompoundButton buttonView, final boolean isChecked) {
-			cDarkText.setEnabled(isChecked);
-			if (!isChecked) {
-				cDarkText.setChecked(false);
-			}
 		}
 	};
 
