@@ -1,13 +1,18 @@
 package yuku.alkitab.base.ac;
 
-import android.app.AlertDialog;
 import android.content.Intent;
 import android.os.Bundle;
 import android.preference.Preference;
+import android.widget.Toast;
+import com.afollestad.materialdialogs.AlertDialogWrapper;
+import gnu.trove.set.TLongSet;
+import yuku.afw.storage.Preferences;
 import yuku.alkitab.base.App;
 import yuku.alkitab.base.S;
 import yuku.alkitab.base.ac.base.BasePreferenceActivity;
 import yuku.alkitab.base.model.MVersionDb;
+import yuku.alkitab.base.storage.Prefkey;
+import yuku.alkitab.base.util.Announce;
 import yuku.alkitab.base.util.Sqlitil;
 import yuku.alkitab.debug.R;
 import yuku.alkitab.model.ProgressMark;
@@ -28,7 +33,7 @@ public class SecretSettingsActivity extends BasePreferenceActivity {
 			labels[i] = progressMark.caption + " (preset_id " + progressMark.preset_id + ")";
 		}
 
-		new AlertDialog.Builder(SecretSettingsActivity.this)
+		new AlertDialogWrapper.Builder(SecretSettingsActivity.this)
 			.setItems(labels, (dialog, which) -> {
 				final List<ProgressMarkHistory> pmhs = S.getDb().listProgressMarkHistoryByPresetId(progressMarks.get(which).preset_id);
 				final String[] items = new String[pmhs.size()];
@@ -36,7 +41,7 @@ public class SecretSettingsActivity extends BasePreferenceActivity {
 					final ProgressMarkHistory pmh = pmhs.get(i);
 					items[i] = "'" + pmh.progress_mark_caption + "' " + Sqlitil.toLocaleDateMedium(pmh.createTime) + ": " + S.activeVersion.reference(pmh.ari);
 				}
-				new AlertDialog.Builder(SecretSettingsActivity.this)
+				new AlertDialogWrapper.Builder(SecretSettingsActivity.this)
 					.setItems(items, null)
 					.show();
 			})
@@ -55,7 +60,7 @@ public class SecretSettingsActivity extends BasePreferenceActivity {
 			);
 		}
 
-		new AlertDialog.Builder(SecretSettingsActivity.this)
+		new AlertDialogWrapper.Builder(SecretSettingsActivity.this)
 			.setItems(items.toArray(new String[items.size()]), null)
 			.show();
 
@@ -72,6 +77,14 @@ public class SecretSettingsActivity extends BasePreferenceActivity {
 		return true;
 	};
 
+	Preference.OnPreferenceClickListener secret_reset_read_announcements = preference -> {
+		final TLongSet read = Announce.getReadAnnouncementIds();
+		Preferences.remove(Prefkey.announce_read_ids);
+
+		Toast.makeText(this, "Cleared read announcement ids.\n\nPreviously has " + read.size() + " items:\n" + read, Toast.LENGTH_LONG).show();
+		return true;
+	};
+
 	@SuppressWarnings("deprecation")
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
@@ -83,6 +96,7 @@ public class SecretSettingsActivity extends BasePreferenceActivity {
 		findPreference("secret_version_table").setOnPreferenceClickListener(secret_version_table_click);
 		findPreference("secret_sync_debug").setOnPreferenceClickListener(secret_sync_debug);
 		findPreference("secret_old_v3_marker_import").setOnPreferenceClickListener(secret_old_v3_marker_import);
+		findPreference("secret_reset_read_announcements").setOnPreferenceClickListener(secret_reset_read_announcements);
 	}
 
 	public static Intent createIntent() {
