@@ -1,6 +1,5 @@
 package yuku.alkitab.base.ac;
 
-import android.app.AlertDialog;
 import android.content.Intent;
 import android.content.res.Resources;
 import android.os.Bundle;
@@ -10,6 +9,7 @@ import android.preference.ListPreference;
 import android.preference.Preference;
 import android.preference.PreferenceFragment;
 import android.view.View;
+import com.afollestad.materialdialogs.AlertDialogWrapper;
 import yuku.afw.storage.Preferences;
 import yuku.alkitab.base.App;
 import yuku.alkitab.base.ac.base.BasePreferenceActivity;
@@ -23,7 +23,8 @@ import static yuku.alkitab.base.util.Literals.List;
 public class SettingsActivity extends BasePreferenceActivity {
 	public List<String> VALID_FRAGMENT_NAMES = List(
 		DisplayFragment.class.getName(),
-		UsageFragment.class.getName()
+		UsageFragment.class.getName(),
+		CopyShareFragment.class.getName()
 	);
 	Header firstHeaderWithFragment;
 
@@ -102,7 +103,7 @@ public class SettingsActivity extends BasePreferenceActivity {
 				final boolean value = (boolean) newValue;
 
 				if (value) {
-					new AlertDialog.Builder(getActivity())
+					new AlertDialogWrapper.Builder(getActivity())
 						.setMessage(R.string.show_hidden_version_warning)
 						.setNegativeButton(R.string.cancel, null)
 						.setPositiveButton(R.string.ok, (dialog, which) -> pref_showHiddenVersion.setChecked(true))
@@ -115,6 +116,15 @@ public class SettingsActivity extends BasePreferenceActivity {
 		}
 	}
 
+	public static class CopyShareFragment extends PreferenceFragment {
+		@Override
+		public void onCreate(final Bundle savedInstanceState) {
+			super.onCreate(savedInstanceState);
+
+			addPreferencesFromResource(R.xml.settings_copy_share);
+		}
+	}
+
 	static void autoDisplayListPreference(final ListPreference pref) {
 		final CharSequence label = pref.getEntry();
 		if (label != null) {
@@ -122,26 +132,23 @@ public class SettingsActivity extends BasePreferenceActivity {
 		}
 
 		final Preference.OnPreferenceChangeListener originalChangeListener = pref.getOnPreferenceChangeListener();
-		pref.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
-			@Override
-			public boolean onPreferenceChange(final Preference preference, final Object newValue) {
-				final boolean changed;
+		pref.setOnPreferenceChangeListener((preference, newValue) -> {
+			final boolean changed;
 
-				if (originalChangeListener != null) {
-					changed = originalChangeListener.onPreferenceChange(preference, newValue);
-				} else {
-					changed = true;
-				}
-
-				if (changed) {
-					final int index = pref.findIndexOfValue((String) newValue);
-					if (index >= 0) {
-						pref.setSummary(pref.getEntries()[index]);
-					}
-				}
-
-				return changed;
+			if (originalChangeListener != null) {
+				changed = originalChangeListener.onPreferenceChange(preference, newValue);
+			} else {
+				changed = true;
 			}
+
+			if (changed) {
+				final int index = pref.findIndexOfValue((String) newValue);
+				if (index >= 0) {
+					pref.setSummary(pref.getEntries()[index]);
+				}
+			}
+
+			return changed;
 		});
 	}
 
