@@ -1,8 +1,6 @@
 package yuku.alkitab.base.ac;
 
 import android.app.Activity;
-import android.app.AlertDialog;
-import android.app.ProgressDialog;
 import android.content.Intent;
 import android.content.res.Configuration;
 import android.media.MediaPlayer;
@@ -10,6 +8,7 @@ import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
+import android.support.annotation.Nullable;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v4.app.ShareCompat;
 import android.support.v4.widget.DrawerLayout;
@@ -28,6 +27,8 @@ import android.view.Window;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 import android.widget.Toast;
+import com.afollestad.materialdialogs.AlertDialogWrapper;
+import com.afollestad.materialdialogs.MaterialDialog;
 import yuku.afw.V;
 import yuku.afw.storage.Preferences;
 import yuku.alkitab.base.App;
@@ -36,10 +37,12 @@ import yuku.alkitab.base.U;
 import yuku.alkitab.base.ac.base.BaseLeftDrawerActivity;
 import yuku.alkitab.base.dialog.VersesDialog;
 import yuku.alkitab.base.storage.Prefkey;
+import yuku.alkitab.base.storage.SongDb;
 import yuku.alkitab.base.util.AlphanumComparator;
 import yuku.alkitab.base.util.FontManager;
 import yuku.alkitab.base.util.OsisBookNames;
 import yuku.alkitab.base.util.SongBookUtil;
+import yuku.alkitab.base.util.Sqlitil;
 import yuku.alkitab.base.util.TargetDecoder;
 import yuku.alkitab.base.widget.LeftDrawer;
 import yuku.alkitab.base.widget.TwofingerLinearLayout;
@@ -74,6 +77,7 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 	private static final String BIBLE_PROTOCOL = "bible";
 	private static final int REQCODE_songList = 1;
 	private static final int REQCODE_share = 2;
+	private static final int REQCODE_downloadSongBook = 3;
 
 	DrawerLayout drawerLayout;
 	ActionBarDrawerToggle drawerToggle;
@@ -311,7 +315,7 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 						final Activity activity = activityRef.get();
 						if (activity != null) {
 							if (!activity.isFinishing()) {
-								new AlertDialog.Builder(activity)
+								new AlertDialogWrapper.Builder(activity)
 									.setMessage(activity.getString(R.string.song_player_error_description, what, extra))
 									.setPositiveButton(R.string.ok, null)
 									.show();
@@ -390,7 +394,11 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		actionBar.setDisplayHomeAsUpEnabled(true);
 		actionBar.setHomeButtonEnabled(true);
 
-		bDownload.setOnClickListener(v -> SongBookUtil.getSongBookDialog(this, SongBookUtil.getSongBookOnDialogClickListener(this::songBookSelected)).show());
+		bDownload.setOnClickListener(v -> openDownloadSongBookPage());
+	}
+
+	void openDownloadSongBookPage() {
+		startActivityForResult(HelpActivity.createIntent("https://alkitab-host.appspot.com/songs/downloads?app_versionCode=" + App.getVersionCode() + "&app_versionName=" + Uri.encode(App.getVersionName())), REQCODE_downloadSongBook);
 	}
 
 	@Override
@@ -438,9 +446,12 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 			if (bookName == null || code == null) {
 				displaySong(null, null, true);
 			} else {
-				displaySong(bookName, S.getSongDb().getSong(bookName, code), true);
+				final SongDb db = S.getSongDb();
+				displaySong(bookName, db.getSong(bookName, code), true);
 			}
 		}
+
+		getWindow().getDecorView().setKeepScreenOn(Preferences.getBoolean(getString(R.string.pref_keepScreenOn_key), getResources().getBoolean(R.bool.pref_keepScreenOn_default)));
 	}
 
 	@Override
@@ -564,7 +575,7 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		} return true;
 
         case R.id.menuUpdateBook: {
-			new AlertDialog.Builder(this)
+			new AlertDialogWrapper.Builder(this)
 				.setMessage(TextUtils.expandTemplate(getString(R.string.sn_update_book_explanation), currentBookName))
 				.setPositiveButton(R.string.sn_update_book_confirm_button, (dialog, which) -> updateSongBook())
 				.setNegativeButton(R.string.cancel, null)
@@ -572,7 +583,7 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		} return true;
 
 		case R.id.menuDeleteAll: {
-			new AlertDialog.Builder(this)
+			new AlertDialogWrapper.Builder(this)
 				.setMessage(R.string.sn_delete_all_songs_explanation)
 				.setPositiveButton(R.string.delete, (dialog, which) -> deleteAllSongs())
 				.setNegativeButton(R.string.cancel, null)
@@ -590,12 +601,13 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		}
 
 		final String currentSongCode = currentSong.code;
+		final int dataFormatVersion = S.getSongDb().getDataFormatVersionForSongs(currentBookName);
 
-		SongBookUtil.downloadSongBook(SongViewActivity.this, songBookInfo, new SongBookUtil.OnDownloadSongBookListener() {
+		SongBookUtil.downloadSongBook(SongViewActivity.this, songBookInfo, dataFormatVersion, new SongBookUtil.OnDownloadSongBookListener() {
 			@Override
 			public void onFailedOrCancelled(SongBookUtil.SongBookInfo songBookInfo, Exception e) {
 				if (e != null) {
-					new AlertDialog.Builder(SongViewActivity.this)
+					new AlertDialogWrapper.Builder(SongViewActivity.this)
 						.setMessage(e.getClass().getSimpleName() + ' ' + e.getMessage())
 						.setPositiveButton(R.string.ok, null)
 						.show();
@@ -604,24 +616,28 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 
 			@Override
 			public void onDownloadedAndInserted(SongBookUtil.SongBookInfo songBookInfo) {
-				final Song song = S.getSongDb().getSong(songBookInfo.bookName, currentSongCode);
-				cache_codes.remove(songBookInfo.bookName);
-				displaySong(songBookInfo.bookName, song);
+				final Song song = S.getSongDb().getSong(songBookInfo.name, currentSongCode);
+				cache_codes.remove(songBookInfo.name);
+				displaySong(songBookInfo.name, song);
 			}
 		});
 	}
 
     protected void deleteAllSongs() {
-        final ProgressDialog pd = ProgressDialog.show(this, null, getString(R.string.please_wait_titik3), true, false);
+		final MaterialDialog pd = new MaterialDialog.Builder(this)
+			.content(R.string.please_wait_titik3)
+			.cancelable(false)
+			.progress(true, 0)
+			.show();
 
-        new Thread() {
+		new Thread() {
             @Override public void run() {
                 final int count = S.getSongDb().deleteAllSongs();
 
                 runOnUiThread(() -> {
                     pd.dismiss();
 
-					new AlertDialog.Builder(SongViewActivity.this)
+					new AlertDialogWrapper.Builder(SongViewActivity.this)
 						.setMessage(getString(R.string.sn_delete_all_songs_result, count))
 						.setPositiveButton(R.string.ok, null)
 						.show()
@@ -806,11 +822,11 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		return res;
 	}
 
-	void displaySong(String bookName, Song song) {
+	void displaySong(String bookName, @Nullable Song song) {
 		displaySong(bookName, song, false);
 	}
 
-	void displaySong(String bookName, Song song, boolean onCreate) {
+	void displaySong(String bookName, @Nullable Song song, boolean onCreate) {
 		song_container.setVisibility(song != null? View.VISIBLE: View.GONE);
 		no_song_data_container.setVisibility(song != null? View.GONE: View.VISIBLE);
 
@@ -829,7 +845,8 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		// construct rendition of scripture references
 		String scripture_references = renderScriptureReferences(BIBLE_PROTOCOL, song.scriptureReferences);
 		templateCustomVars.putString("scripture_references", scripture_references);
-		templateCustomVars.putString("copyright", SongBookUtil.getCopyright(bookName));
+		final String copyright = SongBookUtil.getCopyright(bookName);
+		templateCustomVars.putString("copyright", copyright != null? copyright: "");
 
 		FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
 		ft.replace(R.id.song_container, SongFragment.create(song, "templates/song.html", templateCustomVars));
@@ -862,24 +879,94 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		state_tempCode = "";
 	}
 
-	@Override protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-		if (requestCode == REQCODE_songList) {
-			if (resultCode == RESULT_OK) {
-				SongListActivity.Result result = SongListActivity.obtainResult(data);
-				if (result != null) {
-					displaySong(result.bookName, S.getSongDb().getSong(result.bookName, result.code));
-					// store this for next search
-					last_searchState = result.last_searchState;
+	@Override
+	protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+		switch (requestCode) {
+			case REQCODE_songList: {
+				if (resultCode == RESULT_OK) {
+					SongListActivity.Result result = SongListActivity.obtainResult(data);
+					if (result != null) {
+						displaySong(result.bookName, S.getSongDb().getSong(result.bookName, result.code));
+						// store this for next search
+						last_searchState = result.last_searchState;
+					}
 				}
-			}
-		} else if (requestCode == REQCODE_share) {
-			if (resultCode == RESULT_OK) {
-				ShareActivity.Result result = ShareActivity.obtainResult(data);
-				if (result != null && result.chosenIntent != null) {
-					startActivity(result.chosenIntent);
+			} return;
+			case REQCODE_share: {
+				if (resultCode == RESULT_OK) {
+					ShareActivity.Result result = ShareActivity.obtainResult(data);
+					if (result != null && result.chosenIntent != null) {
+						startActivity(result.chosenIntent);
+					}
 				}
-			}
+			} return;
+			case REQCODE_downloadSongBook: {
+				if (resultCode == RESULT_OK) {
+					final Uri uri = data.getData();
+					if (uri != null) {
+						downloadByAlkitabUri(uri);
+					}
+					return;
+				}
+			} return;
 		}
+
+		super.onActivityResult(requestCode, resultCode, data);
+	}
+
+
+	private void downloadByAlkitabUri(final Uri uri) {
+		if (!"alkitab".equals(uri.getScheme()) || !"/addon/download".equals(uri.getPath()) || !"songbook".equals(uri.getQueryParameter("kind")) || !"ser".equals(uri.getQueryParameter("type")) || uri.getQueryParameter("name") == null) {
+			new MaterialDialog.Builder(this)
+				.content("Invalid uri:\n\n" + uri)
+				.positiveText(R.string.ok)
+				.show();
+			return;
+		}
+
+		final String dataFormatVersion_s = uri.getQueryParameter("dataFormatVersion");
+		final int dataFormatVersion;
+		try {
+			dataFormatVersion = Integer.parseInt(dataFormatVersion_s);
+		} catch (NumberFormatException|NullPointerException e) {
+			new MaterialDialog.Builder(this)
+				.content("Invalid uri:\n\n" + uri)
+				.positiveText(R.string.ok)
+				.show();
+			return;
+		}
+
+		if (!SongBookUtil.isSupportedDataFormatVersion(dataFormatVersion)) {
+			new MaterialDialog.Builder(this)
+				.content("Unsupported data format version: " + dataFormatVersion)
+				.positiveText(R.string.ok)
+				.show();
+			return;
+		}
+
+		final SongBookUtil.SongBookInfo info = new SongBookUtil.SongBookInfo();
+		info.name = uri.getQueryParameter("name");
+		info.title = uri.getQueryParameter("title");
+		info.copyright = uri.getQueryParameter("copyright");
+
+		SongBookUtil.downloadSongBook(this, info, dataFormatVersion, new SongBookUtil.OnDownloadSongBookListener() {
+			@Override
+			public void onDownloadedAndInserted(final SongBookUtil.SongBookInfo songBookInfo) {
+				final String name = songBookInfo.name;
+				final Song song = S.getSongDb().getFirstSongFromBook(name);
+				displaySong(name, song);
+			}
+
+			@Override
+			public void onFailedOrCancelled(final SongBookUtil.SongBookInfo songBookInfo, final Exception e) {
+				if (e != null) {
+					new AlertDialogWrapper.Builder(SongViewActivity.this)
+						.setMessage(e.getClass().getSimpleName() + ' ' + e.getMessage())
+						.setPositiveButton(R.string.ok, null)
+						.show();
+				}
+			}
+		});
 	}
 
 	static class PatchTextExtraInfoJson {
@@ -904,15 +991,31 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		if ("patchtext".equals(scheme)) {
 			final Song song = currentSong;
 
-			final PatchTextExtraInfoJson extraInfo = new PatchTextExtraInfoJson();
-			extraInfo.type = "song";
-			extraInfo.bookName = currentBookName;
-			extraInfo.code = song.code;
+			// do not proceed if the song is too old
+			final int updateTime = S.getSongDb().getSongUpdateTime(currentBookName, song.code);
+			if (updateTime == 0 || Sqlitil.nowDateTime() - updateTime > 21 * 86400) {
+				new MaterialDialog.Builder(this)
+					.content(TextUtils.expandTemplate(getString(R.string.sn_update_book_because_too_old), currentBookName))
+					.positiveText(R.string.sn_update_book_confirm_button)
+					.negativeText(R.string.cancel)
+					.callback(new MaterialDialog.ButtonCallback() {
+						@Override
+						public void onPositive(final MaterialDialog dialog) {
+							updateSongBook();
+						}
+					})
+					.show();
+			} else {
+				final PatchTextExtraInfoJson extraInfo = new PatchTextExtraInfoJson();
+				extraInfo.type = "song";
+				extraInfo.bookName = currentBookName;
+				extraInfo.code = song.code;
 
-			final String songHeader = song.code + " " + song.title + nonullbr(song.title_original) + nonullbr(song.tune) + nonullbr(song.keySignature) + nonullbr(song.timeSignature) + nonullbr(song.authors_lyric) + nonullbr(song.authors_music);
-			final String songHtml = SongFragment.songToHtml(song, true);
-			final Spanned baseBody = Html.fromHtml(songHeader + "\n\n" + songHtml);
-			startActivity(PatchTextActivity.createIntent(baseBody, App.getDefaultGson().toJson(extraInfo), null));
+				final String songHeader = song.code + " " + song.title + nonullbr(song.title_original) + nonullbr(song.tune) + nonullbr(song.keySignature) + nonullbr(song.timeSignature) + nonullbr(song.authors_lyric) + nonullbr(song.authors_music);
+				final String songHtml = SongFragment.songToHtml(song, true);
+				final Spanned baseBody = Html.fromHtml(songHeader + "\n\n" + songHtml);
+				startActivity(PatchTextActivity.createIntent(baseBody, App.getDefaultGson().toJson(extraInfo), null));
+			}
 			return true;
 		} else if (BIBLE_PROTOCOL.equals(scheme)) {
 			final IntArrayList ariRanges = TargetDecoder.decode("o:" + uri.getSchemeSpecificPart());
@@ -946,23 +1049,30 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 		final int id = v.getId();
 		int num = -1;
 		for (int i = 0; i < numIds.length; i++) if (id == numIds[i]) num = i;
-		for (int i = 0; i < alphaIds.length; i++) if (id == alphaIds[i]) num = 10 + i;
+		for (int i = 0; i < alphaIds.length; i++) if (id == alphaIds[i]) num = 10 + i; // special code for alpha
+		if (id == R.id.bBackspace) num = 20; // special code for backspace
 
 		final LeftDrawer.Songs.Handle handle = leftDrawer.getHandle();
 
-		if (num >= 0) { // digits or letters
-			if (state_tempCode.length() >= 4) state_tempCode = ""; // can't be more than 4 digits
-
-			if (num <= 9) { // digits
-				//noinspection StatementWithEmptyBody
-				if (state_tempCode.length() == 0 && num == 0) { // nothing has been pressed and 0 is now pressed
-				} else {
-					state_tempCode += num;
+		if (num >= 0) { // digits or letters or backspace
+			if (num == 20) { // backspace
+				if (state_tempCode.length() > 0) {
+					state_tempCode = state_tempCode.substring(0, state_tempCode.length() - 1);
 				}
-			} else { // letters
-				final char letter = (char) ('A' + num - 10);
-				if (state_tempCode.length() != 0) {
-					state_tempCode += letter;
+			} else {
+				if (state_tempCode.length() >= 4) state_tempCode = ""; // can't be more than 4 digits
+
+				if (num <= 9) { // digits
+					if (state_tempCode.length() == 0 && num == 0) {
+						// nothing has been pressed and 0 is now pressed
+					} else {
+						state_tempCode += num;
+					}
+				} else if (num <= 19) { // letters
+					final char letter = (char) ('A' + num - 10);
+					if (state_tempCode.length() != 0) {
+						state_tempCode += letter;
+					}
 				}
 			}
 
@@ -988,36 +1098,23 @@ public class SongViewActivity extends BaseLeftDrawerActivity implements SongFrag
 	}
 
 	@Override
-	public void songBookSelected(final boolean all, final SongBookUtil.SongBookInfo songBookInfo) {
-		if (all) return; // should not happen
-
-		final Song song = S.getSongDb().getFirstSongFromBook(songBookInfo.bookName);
+	public void songBookSelected(final String name) {
+		final Song song = S.getSongDb().getFirstSongFromBook(name);
 
 		if (song != null) {
-			displaySong(songBookInfo.bookName, song);
-		} else {
-			SongBookUtil.downloadSongBook(SongViewActivity.this, songBookInfo, new SongBookUtil.OnDownloadSongBookListener() {
-				@Override public void onFailedOrCancelled(SongBookUtil.SongBookInfo songBookInfo, Exception e) {
-					if (e != null) {
-						new AlertDialog.Builder(SongViewActivity.this)
-							.setMessage(e.getClass().getSimpleName() + ' ' + e.getMessage())
-							.setPositiveButton(R.string.ok, null)
-							.show();
-					}
-				}
-
-				@Override public void onDownloadedAndInserted(SongBookUtil.SongBookInfo songBookInfo) {
-					Song song = S.getSongDb().getFirstSongFromBook(songBookInfo.bookName);
-					displaySong(songBookInfo.bookName, song);
-				}
-			});
+			displaySong(name, song);
 		}
 
 		state_tempCode = "";
 	}
 
+	@Override
+	public void moreSelected() {
+		openDownloadSongBookPage();
+	}
+
 	void setCustomProgressBarIndeterminateVisible(final boolean visible) {
-		circular_progress.setVisibility(visible ? View.VISIBLE : View.INVISIBLE);
+		circular_progress.setVisibility(visible ? View.VISIBLE : View.GONE);
 	}
 }
 
