@@ -1,13 +1,13 @@
 package yuku.alkitab.base.devotion;
 
 import android.net.Uri;
+import android.support.annotation.NonNull;
 import android.text.Html;
 import android.text.SpannableStringBuilder;
-import android.text.style.URLSpan;
 import yuku.alkitab.base.ac.DevotionActivity;
 import yuku.alkitab.base.widget.CallbackSpan;
 
-public class ArticleMeidA implements DevotionArticle {
+public class ArticleMeidA extends DevotionArticle {
 	public static final String TAG = ArticleMeidA.class.getSimpleName();
 	private String date;
 	private String bodyHtml;
@@ -23,11 +23,13 @@ public class ArticleMeidA implements DevotionArticle {
 		this.readyToUse = readyToUse;
 	}
 
-	@Override public String getDate() {
+	@Override
+	public String getDate() {
 		return date;
 	}
 
-	@Override public boolean getReadyToUse() {
+	@Override
+	public boolean getReadyToUse() {
 		return readyToUse;
 	}
 
@@ -36,32 +38,25 @@ public class ArticleMeidA implements DevotionArticle {
 		return DevotionActivity.DevotionKind.MEID_A;
 	}
 
-	@Override public String getRawEncoding() {
-		return "utf-8";
-	}
-
-	@Override public void fillIn(String raw) {
+	@Override
+	public void fillIn(String raw) {
 		bodyHtml = raw;
 		readyToUse = !raw.startsWith("NG");
 	}
 
-	@Override public CharSequence getContent(CallbackSpan.OnClickListener<String> verseClickListener) {
+	@Override
+	public CharSequence getContent(CallbackSpan.OnClickListener<String> verseClickListener) {
 		SpannableStringBuilder sb = new SpannableStringBuilder();
 		sb.append(Html.fromHtml(bodyHtml + "<br/><small><a href='patchtext://host/?referenceUrl=" + Uri.encode("http://m.ccel.org/ccel/spurgeon/morneve.d" + date.substring(4, 8) + "am.html") + "'>Kirim perbaikan</a></small>"));
-		
-		// replace URLSpans with CallbackSpans
-		URLSpan[] spans = sb.getSpans(0, sb.length(), URLSpan.class);
-		for (URLSpan oldSpan: spans) {
-			String url = oldSpan.getURL();
-			CallbackSpan<String> newSpan = new CallbackSpan<>(url, verseClickListener);
-			sb.setSpan(newSpan, sb.getSpanStart(oldSpan), sb.getSpanEnd(oldSpan), 0);
-			sb.removeSpan(oldSpan);
-		}
-		
+
+		convertLinks(sb, verseClickListener);
+
 		return sb;
 	}
 
-	@Override public String[] getHeaderTitleBody() {
-		return new String[] {null, null, bodyHtml};
+	@Override
+	@NonNull
+	public String getBody() {
+		return bodyHtml;
 	}
 }
