@@ -1,33 +1,34 @@
 package yuku.alkitab.base.devotion;
 
+import android.support.annotation.NonNull;
 import android.text.Html;
 import android.text.SpannableStringBuilder;
-import android.text.style.URLSpan;
-
 import yuku.alkitab.base.ac.DevotionActivity;
 import yuku.alkitab.base.widget.CallbackSpan;
 
-public class ArticleMorningEveningEnglish implements DevotionArticle {
+public class ArticleMorningEveningEnglish extends DevotionArticle {
 	public static final String TAG = ArticleMorningEveningEnglish.class.getSimpleName();
 	private String date;
-	private String bodyHtml;
+	private String body;
 	private boolean readyToUse;
-	
+
 	public ArticleMorningEveningEnglish(String date) {
 		this.date = date;
 	}
 
-	public ArticleMorningEveningEnglish(String date, String bodyHtml, boolean readyToUse) {
+	public ArticleMorningEveningEnglish(String date, String body, boolean readyToUse) {
 		this.date = date;
-		this.bodyHtml = bodyHtml;
+		this.body = body;
 		this.readyToUse = readyToUse;
 	}
 
-	@Override public String getDate() {
+	@Override
+	public String getDate() {
 		return date;
 	}
 
-	@Override public boolean getReadyToUse() {
+	@Override
+	public boolean getReadyToUse() {
 		return readyToUse;
 	}
 
@@ -36,32 +37,25 @@ public class ArticleMorningEveningEnglish implements DevotionArticle {
 		return DevotionActivity.DevotionKind.ME_EN;
 	}
 
-	@Override public String getRawEncoding() {
-		return "utf-8";
-	}
-
-	@Override public void fillIn(String raw) {
-		bodyHtml = raw;
+	@Override
+	public void fillIn(String raw) {
+		body = raw;
 		readyToUse = !raw.startsWith("NG");
 	}
 
-	@Override public CharSequence getContent(CallbackSpan.OnClickListener<String> verseClickListener) {
+	@Override
+	public CharSequence getContent(CallbackSpan.OnClickListener<String> verseClickListener) {
 		SpannableStringBuilder sb = new SpannableStringBuilder();
-		sb.append(Html.fromHtml(bodyHtml));
-		
-		// replace URLSpans with CallbackSpans
-		URLSpan[] spans = sb.getSpans(0, sb.length(), URLSpan.class);
-		for (URLSpan oldSpan: spans) {
-			String url = oldSpan.getURL();
-			CallbackSpan<String> newSpan = new CallbackSpan<>(url, verseClickListener);
-			sb.setSpan(newSpan, sb.getSpanStart(oldSpan), sb.getSpanEnd(oldSpan), 0);
-			sb.removeSpan(oldSpan);
-		}
-		
+		sb.append(Html.fromHtml(body));
+
+		convertLinks(sb, verseClickListener);
+
 		return sb;
 	}
 
-	@Override public String[] getHeaderTitleBody() {
-		return new String[] {null, null, bodyHtml};
+	@Override
+	@NonNull
+	public String getBody() {
+		return body;
 	}
 }
