@@ -1,10 +1,7 @@
 package yuku.alkitab.base.pdbconvert;
 
-import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemSelectedListener;
@@ -12,6 +9,7 @@ import android.widget.ArrayAdapter;
 import android.widget.CheckBox;
 import android.widget.Spinner;
 import android.widget.TextView;
+import com.afollestad.materialdialogs.MaterialDialog;
 import com.compactbyte.android.bible.PDBFileStream;
 import com.compactbyte.bibleplus.reader.BiblePlusPDB;
 import com.compactbyte.bibleplus.reader.BookInfo;
@@ -30,7 +28,7 @@ public class ConvertOptionsDialog {
 	public static final String TAG = ConvertOptionsDialog.class.getSimpleName();
 	
 	Context context;
-	AlertDialog alert;
+	MaterialDialog alert;
 	ConvertOptionsCallback callback;
 	
 	Spinner cbEncoding;
@@ -89,19 +87,20 @@ public class ConvertOptionsDialog {
 			return;
 		}
 
-		View dialogLayout = LayoutInflater.from(context).inflate(R.layout.dialog_pdbconvert_options, null);
-		
-		this.alert = new AlertDialog.Builder(context)
-		.setView(dialogLayout)
-		.setTitle(R.string.pdb_file_options)
-		.setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-			@Override
-			public void onClick(DialogInterface dialog, int which) {
-				bOk_click();
-			}
-		})
-		.setNegativeButton(R.string.cancel, null)
-		.create();
+		this.alert = new MaterialDialog.Builder(context)
+			.customView(R.layout.dialog_pdbconvert_options, false)
+			.title(R.string.pdb_file_options)
+			.positiveText(R.string.ok)
+			.negativeText(R.string.cancel)
+			.callback(new MaterialDialog.ButtonCallback() {
+				@Override
+				public void onPositive(final MaterialDialog dialog) {
+					bOk_click();
+				}
+			})
+			.build();
+
+		final View dialogLayout = this.alert.getCustomView();
 
 		cbEncoding = V.get(dialogLayout, R.id.cbEncoding);
 		lSample = V.get(dialogLayout, R.id.lSample);
