@@ -9,6 +9,7 @@ import android.provider.BaseColumns;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.util.Log;
+import android.util.Pair;
 import com.google.gson.reflect.TypeToken;
 import yuku.afw.D;
 import yuku.afw.storage.Preferences;
@@ -41,6 +42,7 @@ import yuku.alkitab.util.Ari;
 import yuku.alkitab.util.IntArrayList;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
@@ -416,32 +418,27 @@ public class InternalDb {
 	}
 
 	public void storeArticleToDevotions(DevotionArticle article) {
-		SQLiteDatabase db = helper.getWritableDatabase();
+		final SQLiteDatabase db = helper.getWritableDatabase();
+
+		final ContentValues values = new ContentValues();
+		values.put(Table.Devotion.name.name(), article.getKind().name);
+		values.put(Table.Devotion.date.name(), article.getDate());
+		values.put(Table.Devotion.readyToUse.name(), article.getReadyToUse() ? 1 : 0);
+
+		if (article.getReadyToUse()) {
+			values.put(Table.Devotion.body.name(), article.getBody());
+		} else {
+			values.putNull(Table.Devotion.body.name());
+		}
+
+		values.put(Table.Devotion.touchTime.name(), Sqlitil.nowDateTime());
+		values.put(Table.Devotion.dataFormatVersion.name(), 1);
 
 		db.beginTransaction();
 		try {
 			// first delete the existing
-			db.delete(Db.TABLE_Devotion, Db.Devotion.name + "=? and " + Db.Devotion.date + "=?", new String[] {article.getKind().name, article.getDate()});
-
-			ContentValues values = new ContentValues();
-			values.put(Db.Devotion.name, article.getKind().name);
-			values.put(Db.Devotion.date, article.getDate());
-			values.put(Db.Devotion.readyToUse, article.getReadyToUse()? 1: 0);
-
-			if (article.getReadyToUse()) {
-				String[] headerTitleBody = article.getHeaderTitleBody();
-				values.put(Db.Devotion.header, headerTitleBody[0]);
-				values.put(Db.Devotion.title, headerTitleBody[1]);
-				values.put(Db.Devotion.body, headerTitleBody[2]);
-			} else {
-				values.putNull(Db.Devotion.header);
-				values.putNull(Db.Devotion.title);
-				values.putNull(Db.Devotion.body);
-			}
-
-			values.put(Db.Devotion.touchTime, Sqlitil.nowDateTime());
-
-			db.insert(Db.TABLE_Devotion, null, values);
+			db.delete(Table.Devotion.tableName(), Table.Devotion.name + "=? and " + Table.Devotion.date + "=?", new String[]{article.getKind().name, article.getDate()});
+			db.insert(Table.Devotion.tableName(), null, values);
 
 			db.setTransactionSuccessful();
 		} finally {
@@ -450,20 +447,18 @@ public class InternalDb {
 	}
 
 	public int deleteDevotionsWithTouchTimeBefore(Date date) {
-		SQLiteDatabase db = helper.getWritableDatabase();
-		return db.delete(Db.TABLE_Devotion, Db.Devotion.touchTime + "<?", new String[] {String.valueOf(Sqlitil.toInt(date))});
+		final SQLiteDatabase db = helper.getWritableDatabase();
+		return db.delete(Table.Devotion.tableName(), Table.Devotion.touchTime + "<?", ToStringArray(Sqlitil.toInt(date)));
 	}
 
 	/**
 	 * Try to get article from local db. Non ready-to-use article will be returned too.
 	 */
 	public DevotionArticle tryGetDevotion(String name, String date) {
-		Cursor c = helper.getReadableDatabase().query(Db.TABLE_Devotion, null, Db.Devotion.name + "=? and " + Db.Devotion.date + "=?", new String[]{name, date}, null, null, null);
+		final Cursor c = helper.getReadableDatabase().query(Table.Devotion.tableName(), null, Table.Devotion.name + "=? and " + Table.Devotion.date + "=? and " + Table.Devotion.dataFormatVersion + "=?", ToStringArray(name, date, 1), null, null, null);
 		try {
-			int col_title = c.getColumnIndexOrThrow(Db.Devotion.title);
-			int col_header = c.getColumnIndexOrThrow(Db.Devotion.header);
-			int col_body = c.getColumnIndexOrThrow(Db.Devotion.body);
-			int col_readyToUse = c.getColumnIndexOrThrow(Db.Devotion.readyToUse);
+			final int col_body = c.getColumnIndexOrThrow(Table.Devotion.body.name());
+			final int col_readyToUse = c.getColumnIndexOrThrow(Table.Devotion.readyToUse.name());
 
 			if (!c.moveToNext()) {
 				return null;
@@ -472,22 +467,10 @@ public class InternalDb {
 			final DevotionActivity.DevotionKind kind = DevotionActivity.DevotionKind.getByName(name);
 			switch (kind) {
 				case RH: {
-					return new ArticleRenunganHarian(
-					date,
-					c.getString(col_title),
-					c.getString(col_header),
-					c.getString(col_body),
-					c.getInt(col_readyToUse) > 0
-					);
+					return new ArticleRenunganHarian(date, c.getString(col_body), c.getInt(col_readyToUse) > 0);
 				}
 				case SH: {
-					return new ArticleSantapanHarian(
-					date,
-					c.getString(col_title),
-					c.getString(col_header),
-					c.getString(col_body),
-					c.getInt(col_readyToUse) > 0
-					);
+					return new ArticleSantapanHarian(date, c.getString(col_body), c.getInt(col_readyToUse) > 0);
 				}
 				case ME_EN: {
 					return new ArticleMorningEveningEnglish(date, c.getString(col_body), true);
@@ -883,9 +866,30 @@ public class InternalDb {
 		return (int) DatabaseUtils.longForQuery(db, "select count(*) from " + Db.TABLE_Marker_Label + " where " + Db.Marker_Label.label_gid + "=?", new String[]{label.gid});
 	}
 
-	public int deleteDevotionsWithLessThanInTitle() {
-		SQLiteDatabase db = helper.getWritableDatabase();
-		return db.delete(Db.TABLE_Devotion, Db.Devotion.title + " like '%<%'", null);
+	public void sortLabelsAlphabetically() {
+		final SQLiteDatabase db = helper.getWritableDatabase();
+		db.beginTransaction();
+		try {
+			final List<Label> labels = listAllLabels();
+			Collections.sort(labels, (lhs, rhs) -> {
+				if (lhs.title == null || rhs.title == null) {
+					return 0;
+				}
+
+				return lhs.title.compareToIgnoreCase(rhs.title);
+			});
+
+			for (int i = 0; i < labels.size(); i++) {
+				final Label label = labels.get(i);
+				label.ordering = i + 1;
+				db.update(Db.TABLE_Label, labelToContentValues(label), "_id=?", ToStringArray(label._id));
+			}
+
+			db.setTransactionSuccessful();
+		} finally {
+			db.endTransaction();
+		}
+		Sync.notifySyncNeeded(SyncShadow.SYNC_SET_MABEL);
 	}
 
 	public void reorderLabels(Label from, Label to) {
@@ -1134,14 +1138,16 @@ public class InternalDb {
 		return infos;
 	}
 
-	public byte[] getBinaryReadingPlanById(long id) {
-		byte[] buffer = null;
-		final Cursor c = helper.getReadableDatabase().query(Db.TABLE_ReadingPlan, new String[] {Db.ReadingPlan.data}, "_id=?", new String[] {String.valueOf(id)}, null, null, null);
-		while (c.moveToNext()) {
-			buffer = c.getBlob(0);
+	public Pair<String, byte[]> getReadingPlanNameAndData(long _id) {
+		final Cursor c = helper.getReadableDatabase().query(Db.TABLE_ReadingPlan, Array(Db.ReadingPlan.name, Db.ReadingPlan.data), "_id=?", ToStringArray(_id), null, null, null);
+		try {
+			if (c.moveToNext()) {
+				return Pair.create(c.getString(0), c.getBlob(1));
+			}
+			return null;
+		} finally {
+			c.close();
 		}
-		c.close();
-		return buffer;
 	}
 
 	public IntArrayList getAllReadingCodesByReadingPlanId(long id) {
@@ -1179,24 +1185,88 @@ public class InternalDb {
 	}
 
 	@Nullable public SyncShadow getSyncShadowBySyncSetName(final String syncSetName) {
+		// Getting a sync shadow that has a size bigger than 2 MB will cause crash,
+		// because of system CursorWindow implementation that sets the max memory allocated
+		// to be 2 MB, as defined in system resource:
+		// <integer name="config_cursorWindowSize">2048</integer>
+		// So we will get the size first, and then allocate memory,
+		// and get the data in chunks.
 		final SQLiteDatabase db = helper.getReadableDatabase();
-		final Cursor c = db.query(Table.SyncShadow.tableName(), Array(
-			Table.SyncShadow.syncSetName.name(),
-			Table.SyncShadow.revno.name(),
-			Table.SyncShadow.data.name()
-		), Table.SyncShadow.syncSetName + "=?", Array(syncSetName), null, null, null);
+		db.beginTransaction();
 		try {
-			if (c.moveToNext()) {
-				final SyncShadow res = new SyncShadow();
-				res.syncSetName = c.getString(0);
-				res.revno = c.getInt(1);
-				res.data = c.getBlob(2);
-				return res;
+			final int data_len;
+			final long _id;
+			final int revno;
+
+			{ // get blob len
+				final Cursor c = db.rawQuery(
+					"select "
+						+ Table.SyncShadow.revno.name() + ", " // col 0
+						+ "length(" + Table.SyncShadow.data.name() + "), " // col 1
+						+ "_id " // col 2
+						+ " from " + Table.SyncShadow.tableName()
+						+ " where " + Table.SyncShadow.syncSetName + "=?",
+					Array(syncSetName)
+				);
+				try {
+					if (c.moveToNext()) {
+						revno = c.getInt(0);
+						data_len = c.getInt(1);
+						_id = c.getLong(2);
+					} else {
+						return null;
+					}
+				} finally {
+					c.close();
+				}
 			}
+
+			final byte[] data = new byte[data_len];
+
+			{ // fill in blob
+				final int chunkSize = 1000_000;
+				for (int i = 0; i < data_len; i += chunkSize) {
+					final Cursor c = db.rawQuery(
+						// sqlite substr func is 1-indexed
+						"select "
+							+ "substr(" + Table.SyncShadow.data.name() + ", " + (i + 1) + ", " + chunkSize + ")" // col 0
+							+ " from " + Table.SyncShadow.tableName()
+							+ " where _id=?",
+						ToStringArray(_id)
+					);
+
+					try {
+						if (c.moveToNext()) {
+							final byte[] chunk = c.getBlob(0);
+							if (i + chunk.length != data_len) {
+								// not the last one
+								if (chunk.length != chunkSize) {
+									throw new RuntimeException("Not the requested size of chunk retrieved. data_len=" + data_len + " i=" + i + " chunk.len=" + chunk.length);
+								}
+								System.arraycopy(chunk, 0, data, i, chunkSize);
+							} else {
+								// the last one
+								System.arraycopy(chunk, 0, data, i, chunk.length);
+							}
+						} else {
+							throw new RuntimeException("Cursor moveToNext returns false, does not make sense, since previous query has indicated that this cursor has rows.");
+						}
+					} finally {
+						c.close();
+					}
+				}
+			}
+
+			db.setTransactionSuccessful();
+
+			final SyncShadow res = new SyncShadow();
+			res.syncSetName = syncSetName;
+			res.revno = revno;
+			res.data = data;
+			return res;
 		} finally {
-			c.close();
+			db.endTransaction();
 		}
-		return null;
 	}
 
 	@Nullable public int getRevnoFromSyncShadowBySyncSetName(final String syncSetName) {
