@@ -16,7 +16,6 @@
 
 package com.example.android.wizardpager;
 
-import android.app.AlertDialog;
 import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
@@ -29,11 +28,12 @@ import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentStatePagerAdapter;
 import android.support.v4.view.ViewPager;
-import android.util.TypedValue;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.Toast;
+
+import com.afollestad.materialdialogs.AlertDialogWrapper;
 import com.example.android.wizardpager.wizard.model.AbstractWizardModel;
 import com.example.android.wizardpager.wizard.model.CustomerInfoPage;
 import com.example.android.wizardpager.wizard.model.ModelCallbacks;
@@ -42,11 +42,12 @@ import com.example.android.wizardpager.wizard.model.TextareaPage;
 import com.example.android.wizardpager.wizard.ui.PageFragmentCallbacks;
 import com.example.android.wizardpager.wizard.ui.ReviewFragment;
 import com.example.android.wizardpager.wizard.ui.StepPagerStrip;
+
+import java.util.List;
+
 import yuku.alkitabfeedback.AlkitabFeedbackModel;
 import yuku.alkitabfeedback.FeedbackSender;
 import yuku.alkitabfeedback.R;
-
-import java.util.List;
 
 public class MainActivity extends FragmentActivity implements
         PageFragmentCallbacks,
@@ -104,14 +105,14 @@ public class MainActivity extends FragmentActivity implements
 	    mPager.setAdapter(mPagerAdapter);
         mStepPagerStrip = (StepPagerStrip) findViewById(R.id.strip);
         mStepPagerStrip.setOnPageSelectedListener(new StepPagerStrip.OnPageSelectedListener() {
-            @Override
-            public void onPageStripSelected(int position) {
-                position = Math.min(mPagerAdapter.getCount() - 1, position);
-                if (mPager.getCurrentItem() != position) {
-                    mPager.setCurrentItem(position);
-                }
-            }
-        });
+			@Override
+			public void onPageStripSelected(int position) {
+				position = Math.min(mPagerAdapter.getCount() - 1, position);
+				if (mPager.getCurrentItem() != position) {
+					mPager.setCurrentItem(position);
+				}
+			}
+		});
 
         mNextButton = (Button) findViewById(R.id.next_button);
         mPrevButton = (Button) findViewById(R.id.prev_button);
@@ -132,62 +133,64 @@ public class MainActivity extends FragmentActivity implements
         });
 
         mNextButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                if (mPager.getCurrentItem() == mCurrentPageSequence.size()) {
-                    DialogFragment dg = new DialogFragment() {
-                        @Override public Dialog onCreateDialog(Bundle savedInstanceState) {
-                            return new AlertDialog.Builder(getActivity())
-                            .setMessage(R.string.alkitabfeedback_submit_confirm_message)
-                            .setPositiveButton(R.string.alkitabfeedback_submit_confirm_button, new DialogInterface.OnClickListener() {
-								@Override public void onClick(DialogInterface dialog, int which) {
-									String feedback_from_name = null;
-									String feedback_from_email = null;
-									String feedback_body = null;
-									
-									Bundle saved = mWizardModel.save();
-									Bundle contact = saved.getBundle("contact");
-									if (contact != null) {
-										feedback_from_name = contact.getString(CustomerInfoPage.NAME_DATA_KEY);
-										feedback_from_email = contact.getString(CustomerInfoPage.EMAIL_DATA_KEY);
+			@Override
+			public void onClick(final View v) {
+				if (mPager.getCurrentItem() == mCurrentPageSequence.size()) {
+					DialogFragment dg = new DialogFragment() {
+						@Override public Dialog onCreateDialog(Bundle savedInstanceState) {
+							return new AlertDialogWrapper.Builder(getActivity())
+								.setMessage(R.string.alkitabfeedback_submit_confirm_message)
+								.setPositiveButton(R.string.alkitabfeedback_submit_confirm_button, new DialogInterface.OnClickListener() {
+									@Override
+									public void onClick(final DialogInterface dialog, final int which) {
+										String feedback_from_name = null;
+										String feedback_from_email = null;
+										String feedback_body = null;
+
+										Bundle saved = mWizardModel.save();
+										Bundle contact = saved.getBundle("contact");
+										if (contact != null) {
+											feedback_from_name = contact.getString(CustomerInfoPage.NAME_DATA_KEY);
+											feedback_from_email = contact.getString(CustomerInfoPage.EMAIL_DATA_KEY);
+										}
+										Bundle message = saved.getBundle("message");
+										if (message != null) {
+											feedback_body = message.getString(TextareaPage.SIMPLE_DATA_KEY);
+										}
+
+										if (feedback_from_name != null && feedback_from_email != null && feedback_body != null) {
+											FeedbackSender sender = FeedbackSender.getInstance(getApplicationContext());
+											sender.addEntry(feedback_from_name, feedback_from_email, feedback_body);
+											sender.trySend();
+										}
+
+										Toast.makeText(MainActivity.this, R.string.alkitabfeedback_submit_thanks, Toast.LENGTH_LONG).show();
+										setResult(RESULT_OK);
+										finish();
 									}
-									Bundle message = saved.getBundle("message");
-									if (message != null) {
-										feedback_body = message.getString(TextareaPage.SIMPLE_DATA_KEY);
-									}
-									
-									if (feedback_from_name != null && feedback_from_email != null && feedback_body != null) {
-										FeedbackSender sender = FeedbackSender.getInstance(getApplicationContext());
-										sender.addEntry(feedback_from_name, feedback_from_email, feedback_body);
-										sender.trySend();
-									}
-									
-									Toast.makeText(MainActivity.this, R.string.alkitabfeedback_submit_thanks, Toast.LENGTH_LONG).show();
-									setResult(RESULT_OK);
-									finish();
-								}
-							})
-                            .setNegativeButton(android.R.string.cancel, null)
-                            .create();
-                        }
-                    };
-                    dg.show(getSupportFragmentManager(), "place_order_dialog");
-                } else {
-                    if (mEditingAfterReview) {
-                        mPager.setCurrentItem(mPagerAdapter.getCount() - 1);
-                    } else {
-                        mPager.setCurrentItem(mPager.getCurrentItem() + 1);
-                    }
-                }
-            }
-        });
+								})
+								.setNegativeButton(android.R.string.cancel, null)
+								.create();
+						}
+					};
+					dg.show(getSupportFragmentManager(), "place_order_dialog");
+				} else {
+					if (mEditingAfterReview) {
+						mPager.setCurrentItem(mPagerAdapter.getCount() - 1);
+					} else {
+						mPager.setCurrentItem(mPager.getCurrentItem() + 1);
+					}
+				}
+
+			}
+		});
 
         mPrevButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                mPager.setCurrentItem(mPager.getCurrentItem() - 1);
-            }
-        });
+			@Override
+			public void onClick(final View v) {
+				mPager.setCurrentItem(mPager.getCurrentItem() - 1);
+			}
+		});
 
         onPageTreeChanged();
         updateBottomBar();
@@ -215,10 +218,8 @@ public class MainActivity extends FragmentActivity implements
                     : R.string.alkitabfeedback_next);
             
             if (Build.VERSION.SDK_INT >= 11) {
-	            mNextButton.setBackgroundResource(R.drawable.alkitabfeedback_selectable_item_background);
-	            TypedValue v = new TypedValue();
-	            getTheme().resolveAttribute(android.R.attr.textAppearanceMedium, v, true);
-	            mNextButton.setTextAppearance(this, v.resourceId);
+                mNextButton.setTextColor(getResources().getColor(R.color.alkitabfeedback_accent));
+                mNextButton.setBackgroundResource(R.drawable.alkitabfeedback_selectable_item_background);
             }
             
             mNextButton.setEnabled(position != mPagerAdapter.getCutOffPage());
